// apps/api/src/common/error-codes.ts
export enum ErrorCode {
  FILE_REQUIRED = 'FILE_REQUIRED',
  UNSUPPORTED_FORMAT = 'UNSUPPORTED_FORMAT',
  DOCUMENT_NOT_FOUND = 'DOCUMENT_NOT_FOUND',
  INVALID_STATUS_FOR_REPROCESS = 'INVALID_STATUS_FOR_REPROCESS',
  INVALID_STATUS_FOR_REVIEW = 'INVALID_STATUS_FOR_REVIEW',
  INVALID_STATUS_FOR_REJECT = 'INVALID_STATUS_FOR_REJECT',
  INVALID_STATUS_FOR_APPROVE = 'INVALID_STATUS_FOR_APPROVE',
  DOWNLOAD_NOT_AVAILABLE = 'DOWNLOAD_NOT_AVAILABLE',
  CHECKLIST_NOT_AVAILABLE = 'CHECKLIST_NOT_AVAILABLE',
  PHOTO_NOT_FOUND = 'PHOTO_NOT_FOUND',
  PARSING_FAILED = 'PARSING_FAILED',
  PROCESSING_FAILED = 'PROCESSING_FAILED',
  MISSING_FILE_BUFFER = 'MISSING_FILE_BUFFER',
  UNKNOWN_COUNTRY = 'UNKNOWN_COUNTRY',
  UNKNOWN_TNVED_CODE = 'UNKNOWN_TNVED_CODE',
  UNKNOWN_ROW = 'UNKNOWN_ROW',
  INVALID_STATUS_FOR_SET_CODE = 'INVALID_STATUS_FOR_SET_CODE',
  INVALID_STATUS_FOR_CLARIFY = 'INVALID_STATUS_FOR_CLARIFY',
  USER_NOTE_TOO_SHORT = 'USER_NOTE_TOO_SHORT',
  AI_UNAVAILABLE = 'AI_UNAVAILABLE',
  AI_TIMEOUT = 'AI_TIMEOUT',
  TKS_UNAVAILABLE = 'TKS_UNAVAILABLE',
  FILE_CORRUPTED = 'FILE_CORRUPTED',
  FILE_TOO_BIG = 'FILE_TOO_BIG',
  INVALID_FREIGHT = 'INVALID_FREIGHT',
  INVALID_STATUS_FOR_START = 'INVALID_STATUS_FOR_START',
  INVALID_LINK_TOKEN = 'INVALID_LINK_TOKEN',
  MANAGER_NOT_LINKED = 'MANAGER_NOT_LINKED',
  NO_MANAGERS_AVAILABLE = 'NO_MANAGERS_AVAILABLE',
  CLIENT_ALREADY_CLAIMED = 'CLIENT_ALREADY_CLAIMED',
  CLIENT_NOT_ASSIGNED = 'CLIENT_NOT_ASSIGNED',
  INVALID_DEPOSIT_AMOUNT = 'INVALID_DEPOSIT_AMOUNT',
  INVALID_PACKAGE = 'INVALID_PACKAGE',
  TOPUP_NOT_FOUND = 'TOPUP_NOT_FOUND',
  TOPUP_NOT_PENDING = 'TOPUP_NOT_PENDING',
  TOPUP_ALREADY_CONFIRMED = 'TOPUP_ALREADY_CONFIRMED',
  UPLOAD_RATE_LIMITED = 'UPLOAD_RATE_LIMITED',
  AUTH_RATE_LIMITED = 'AUTH_RATE_LIMITED',
}